Move internal stuff to internal
