+++ test/graph/GraphQLTest.ts
@@ -1,6 +1,7 @@
 import "mocha";
 import * as assert from "power-assert";
 import * as GraphQL from "../../src/graph/graphQL";
+import * as internalGraphql from "../../src/internal/graph/graphQL";
 
 describe("GraphQL", () => {
 
@@ -19,18 +20,18 @@ describe("GraphQL", () => {
 
     it("should successfully validate valid query", () => {
         const query = GraphQL.subscriptionFromFile("../graphql/query/repos", __dirname);
-        const errors = GraphQL.validateQuery(query);
+        const errors = internalGraphql.validateQuery(query);
         assert(errors.length === 0);
     });
 
     it("should successfully validate invalid query and return errors", () => {
-        const errors = GraphQL.validateQuery(InvalidQuery);
+        const errors = internalGraphql.validateQuery(InvalidQuery);
         assert(errors.length === 3);
     });
 
     it("should pretty print errors", () => {
-        const errors = GraphQL.validateQuery(InvalidQuery);
-        const message = GraphQL.prettyPrintErrors(errors, InvalidQuery);
+        const errors = internalGraphql.validateQuery(InvalidQuery);
+        const message = internalGraphql.prettyPrintErrors(errors, InvalidQuery);
         assert(message === `Cannot query field "invalid" on type "Repo". [7,17]
                 invalid
 ----------------^
@@ -48,13 +49,13 @@ query Repos($teamId: ID!, $offset: Int!) {
 
     it("should successfully load query from relative path out of project root", () => {
         const query = GraphQL.subscriptionFromFile("../graphql/query/repos", __dirname );
-        const errors = GraphQL.validateQuery(query);
+        const errors = internalGraphql.validateQuery(query);
         assert(errors.length === 0);
     });
 
     it("should successfully load query from relative path", () => {
         const query = GraphQL.subscriptionFromFile("./someOtherQuery.graphql", __dirname);
-        const errors = GraphQL.validateQuery(query);
+        const errors = internalGraphql.validateQuery(query);
         assert(errors.length === 0);
     });
 
@@ -135,7 +136,7 @@ query Repos($teamId: ID!, $offset: Int!) {
             "./someOtherQueryWithTheSameName",
             __dirname,
             {});
-        query = GraphQL.replaceOperationName(query, "BlaBla");
+        query = internalGraphql.replaceOperationName(query, "BlaBla");
         assert.equal(query, ReplaceQuery1);
     });
